Codificacion conexion y validaciones de campos y modificacion de test
