--- Inventario/Inventario.Tests/MODRES_TPME.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Inventario.Tests
-{
-    [TestClass]
-    public class MODRES_TPME
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-
-            var cadenaP = "Producto1";
-            var cadenaC = "12";
-            int resultado = 1;
-            
-        }
-    }
-}
