--- Inventario/Inventario.Tests/MODUS_FeatureModify.feature
@@ -1,7 +1,0 @@
-﻿Feature: MODUS_FeatureModify
-	Para bdd de la historia Modificar Usuario
-
-@mytag
-Scenario: Actualizar usuario
-	When tengo datos en algun campo
-	Then update tabla usuario
