--- Inventario/Inventario.Tests/UsersSteps.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using TechTalk.SpecFlow;
-
-namespace Inventario.Tests
-{
-    [Binding]
-    public class UsersSteps
-    {
-        [Given(@"I have entered (.*) into the calculator")]
-        public void GivenIHaveEnteredIntoTheCalculator(int p0)
-        {
-            ScenarioContext.Current.Pending();
-        }
-        
-        [When(@"I press add")]
-        public void WhenIPressAdd()
-        {
-            ScenarioContext.Current.Pending();
-        }
-        
-        [Then(@"the result should be (.*) on the screen")]
-        public void ThenTheResultShouldBeOnTheScreen(int p0)
-        {
-            ScenarioContext.Current.Pending();
-        }
-    }
-}
